Update Chart.yaml
admin-tls and node-tls will use ca-issuer by default, only selfsignedIssuer depend on ..Values.certificates.issuer.name. Reason: The Job in the indexer will use those certificates to verify to indexer, It was failed before because both certs is not issued from the same CA. Although I think it's not a good idea to use leaf certificates as Root CA, and the job should use Root CA (charts/wazuh/templates/certs/root/rootca.yaml) certificate for the verification, I will keep it as is for now.
Change the certs structure, so that each certificate has it's own directory.
├── certs
│   ├── admin
│   │   └── certificate.yaml
│   ├── component.crp.yaml
│   ├── dashboard
│   │   └── certificate.yaml
│   ├── filebeat
│   │   └── certificate.yaml
│   ├── node
│   │   └── certificate.yaml
│   ├── node.crp.yaml
│   ├── root
│   │   ├── ca-issuer.yaml
│   │   ├── rootca.yaml
│   │   └── selfsigned-issuer.yaml
│   └── rootca.crp.yaml

add extraSpec param to Statefulset, Dashboard, and Manager. This field is less priorized compare to predefined configuration. Use case: I was needed to set dnsPolicy: None, and dnsConfig field in the podSpec to use the Kube DNS instead of node-local-cache in GKE Autopilot, but there's no configuration allowed. So I added this field, to increase configuration flexibility
# example:
dashboard:
  extraSpec:
    deployment: {}
    pod: {}
    container: {}